preflight_check/pyproject.toml
[project]
name = "preflight-check"
version = "0.1.0"
description = "Preflight check for Azure Agentless Scanner deployment"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.20.0",
    "azure-mgmt-authorization>=4.0.0",
    "azure-mgmt-compute>=34.0.0",
    "azure-mgmt-network>=28.1.0",
    "azure-mgmt-subscription>=3.1.1",
    "msgraph-sdk>=1.22.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "rich>=13.9.4",
    "typer>=0.15.1",
]

[project.scripts]
preflight-check = "preflight_check.preflight_check:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.11.1",
]
